[run large tests]
yt:git_commit:970b3b171965f106b4a9358dc94a28df4b5cc830
yt:arcadia-sync:project:yt
FFO-KCUF

git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_4/yt@5763748 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ server/node/exec_agent/slot_manager.cpp
@@ -223,7 +223,7 @@ std::optional<i64> TSlotManager::GetMemoryLimit() const
         : std::nullopt;
 }
 
-std::optional<i64> TSlotManager::GetCpuLimit() const
+std::optional<double> TSlotManager::GetCpuLimit() const
 {
     return JobEnvironment_ && JobEnvironment_->IsEnabled()
        ? JobEnvironment_->GetCpuLimit()
